Moved 'alchemy' module to package top level and renamed types to be more descriptive.
